Move the server's handshake keys to `offboard.protocol`
