+++ positronic/policy/base.py
@@ -4,22 +4,14 @@ from abc import ABC, abstractmethod
 from collections.abc import Callable, Mapping
 from typing import Any, ClassVar
 
-# The inference handshake: a policy reports these through its ``meta``, a remote policy nests the serving
-# half under ``SERVER``, and the harness records the result under ``POLICY_META``. ``TYPE`` names the policy
-# at the top level and the vendor under ``SERVER``, so a reader composes a prefix with a field:
-# f'{SERVER_META}.{TYPE}'.
+# What a policy reports about itself through its ``meta``; a remote policy nests the server's meta under
+# ``SERVER``, and the harness records the result under ``POLICY_META``. ``TYPE`` names the policy at the top
+# level and the vendor under ``SERVER``, so a reader composes a prefix with a field: f'{SERVER_META}.{TYPE}'.
 TYPE = 'type'
-CHECKPOINT_ID = 'checkpoint_id'
 CHECKPOINT_PATH = 'checkpoint_path'
 EXPERIMENT_NAME = 'experiment_name'
 CONFIG_NAME = 'config_name'
-HOST = 'host'
-PORT = 'port'
 SERVER = 'server'
-# The wire half of the handshake: what the server declares for the rig to build and obey.
-LOCAL_STACK = 'local_stack'
-COMPRESS_IMAGES = 'compress_images'
-POSITRONIC_VERSION = 'positronic_version'
 
 POLICY_META = 'inference.policy'
 SERVER_META = f'{POLICY_META}.{SERVER}'
